Added post route for /:id, refined doughnut.js functions after combining it with index.js from public.
