+++ controllers/api/poll-routes.js
@@ -95,6 +95,58 @@ router.post('/', withAuth, (req, res) => {
     });
 });
 
+router.post('/:id', withAuth, (req, res) => {
+  Poll.findOne({
+    where: {
+      id: req.params.id
+    }
+  })
+    .then(dbPollData => {
+      if (!dbPollData) {
+        res.status(404).json({ message: 'No poll found with this id' });
+        return;
+      }
+
+      const pollData = dbPollData.get({ plain: true });
+
+      const hasVoted = pollData.agree_votes.includes(req.session.user_id) || pollData.disagree_votes.includes(req.session.user_id);
+
+      if (hasVoted) {
+        res.status(403).json({ message: 'User has already voted on this poll' });
+        return;
+      }
+
+      const voteType = req.body.vote;
+
+      if (voteType === 'agree') {
+        pollData.agree.votes.push(req.session.user_id);
+      } else if (voteType === 'disagree') {
+        pollData.disagree_votes.push(req.session.user_id);
+      }
+
+      Poll.update({
+        agree_votes: pollData.agree_votes,
+        disagree_votes: pollData.disagree_votes
+      },
+        {
+          where: {
+            id: req.params.id
+          }
+        })
+        .then(dbPollData => {
+          res.json(dbPollData);
+        })
+        .catch(err => {
+          console.log(err);
+          res.status(500).json(err);
+        });
+    })
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
+    });
+});
+
 // UPDATE poll route
 router.put('/:id', withAuth, (req, res) => {
   Poll.update({
