+++ views/partials/poll.handlebars
@@ -15,6 +15,7 @@
         <div class="flex justify-center gap-4 sm:mb-6 lg:gap-6 lg:col-start-2 lg:row-end-6 lg:row-span-6 lg:mb-0">
             <canvas id="{{ id }}-chart"></canvas> {{! line for hosting the poll chart. }}
         </div>
+        <input type="hidden" id="poll-id" value="{{id}}">
         <div
             class=" flex gap-5 mt-2 col-start-1 row-start-3 self-center sm:mt-0 sm:col-start-2 sm:row-start-2 sm:row-span-2 lg:mt-6 lg:col-start-1 lg:row-start-3 lg:row-end-4">
             <button type="submit"
